+++ Mtf/Client/Browser.php
@@ -97,4 +97,14 @@ interface Browser
      * @return string
      */
     public function getUrl();
+
+    /**
+     * @return string
+     */
+    public function getHtmlSource();
+
+    /**
+     * @return string
+     */
+    public function getScreenshotData();
 }