+++ Mtf/Handler/Curl.php
@@ -29,6 +29,13 @@ abstract class Curl implements HandlerInterface
     protected $_configuration;
 
     /**
+     * Mapping values for data.
+     *
+     * @var array
+     */
+    protected $mappingData;
+
+    /**
      * Constructor
      *
      * @constructor
@@ -37,5 +44,27 @@ abstract class Curl implements HandlerInterface
     public function __construct(Config $configuration)
     {
         $this->_configuration = $configuration;
+    }
+
+    /**
+     * Replace mapping data in fixture data
+     *
+     * @param array $data
+     * @return array
+     */
+    protected function replaceMappingData(array $data)
+    {
+        foreach ($data as $key => $value) {
+            if (is_array($value)) {
+                $data[$key] = $this->replaceMappingData($value);
+            } else {
+                if (!isset($this->mappingData[$key])) {
+                    continue;
+                }
+                $data[$key] = isset($this->mappingData[$key][$value]) ? $this->mappingData[$key][$value] : $value;
+            }
+        }
+
+        return $data;
     }
 }