+++ Mtf/System/Event/EventManagerInterface.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * {license_notice}
+ *
+ * @copyright  {copyright}
+ * @license    {license_link}
+ */
+namespace Mtf\System\Event;
+
+interface EventManagerInterface
+{
+    /**
+     * Dispatches event and call all observers attached to it
+     *
+     * @param array $eventTags
+     * @param array $subjects
+     * @return void
+     */
+    public function dispatchEvent(array $eventTags, array $subjects = []);
+}